+++ packages/expo-image/src/Image.tsx
@@ -45,6 +45,7 @@ export class Image extends React.PureComponent<ImageProps> {
   /**
    * Asynchronously checks if an image exists in the disk cache and resolves to
    * the path of the cached image if it does.
+<<<<<<< HEAD
    * @platform ios
    * @return A promise resolving to the path of the cached image. It will resolve
    * to `false` if the image does not exist in the cache. Resolves to `false` on
@@ -52,6 +53,16 @@ export class Image extends React.PureComponent<ImageProps> {
    */
   static async getCachePath(url: string): Promise<string | false> {
     return await ExpoImageModule.getCachePath(url);
+=======
+   * @param source - The source URL to retrieve the cached version's path for
+   * @platform android
+   * @platform ios
+   * @return A promise resolving to the path of the cached image. It will resolve
+   * to `false` if the image does not exist in the cache.
+   */
+  static async getCachePathAsync(source: string): Promise<string | false> {
+    return await ExpoImageModule.getCachePathAsync(source);
+>>>>>>> feature/image-cache-path
   }
 
   render() {